Another refactoring pass, using deque instead now
